+++ modules/portal-api/src/main/java/com/enonic/xp/portal/url/PortalUrlBuilders.java
@@ -1,6 +1,5 @@
 package com.enonic.xp.portal.url;
 
-import com.enonic.wem.api.content.Content;
 import com.enonic.wem.api.content.ContentId;
 import com.enonic.wem.api.content.ContentPath;
 import com.enonic.xp.portal.PortalContext;
@@ -16,8 +15,7 @@ public final class PortalUrlBuilders
 
     private ContentPath getContentPath()
     {
-        final Content content = this.context.getContent();
-        return content != null ? content.getPath() : null;
+        return this.context.getContentPath();
     }
 
     private <T extends PortalUrlBuilder> T defaults( final T builder )
@@ -26,11 +24,6 @@ public final class PortalUrlBuilders
         builder.workspace( this.context.getWorkspace() );
         builder.contentPath( getContentPath() );
         return builder;
-    }
-
-    public GeneralUrlBuilder generalUrl()
-    {
-        return new GeneralUrlBuilder().baseUri( this.context.getBaseUri() );
     }
 
     public AssetUrlBuilder assetUrl()
@@ -64,32 +57,8 @@ public final class PortalUrlBuilders
     }
 
     @Deprecated
-    public GeneralUrlBuilder createUrl( final String path )
-    {
-        return generalUrl().contentPath( path );
-    }
-
-    @Deprecated
-    public AssetUrlBuilder createResourceUrl( final String path )
-    {
-        return assetUrl().path( path );
-    }
-
-    @Deprecated
-    public ImageUrlBuilder createImageUrl( final String name )
-    {
-        return imageUrl().imageName( name );
-    }
-
-    @Deprecated
     public ImageUrlBuilder createImageByIdUrl( final ContentId contentId )
     {
         return imageUrl().imageId( contentId );
-    }
-
-    @Deprecated
-    public ServiceUrlBuilder createServiceUrl( final String name )
-    {
-        return serviceUrl().service( name );
     }
 }