+++ modules/portal-api/src/main/java/com/enonic/xp/portal/url/PortalUrlService.java
@@ -1,0 +1,8 @@
+package com.enonic.xp.portal.url;
+
+public interface PortalUrlService
+{
+    public String assetUrl( AssetUrlParams params );
+
+    public String serviceUrl( ServiceUrlParams params );
+}