+++ modules/portal-jslib/src/test/resources/com/enonic/xp/portal/jslib/content/ModifyContentHandlerTest.js
@@ -7,16 +7,16 @@ var expectedJson = {
     "_name": "mycontent",
     "_path": "/a/b/mycontent",
     "data": {
-        "a": [2.0],
-        "b": ["2"],
+        "a": 2.0,
+        "b": "2",
         "c": [{
-            "d": ["true"]
+            "d": true
         }, {
-            "d": ["true"],
+            "d": true,
             "e": ["3", "4", "5"],
-            "f": ["2"]
+            "f": 2
         }],
-        "z": ["99"]
+        "z": "99"
     },
     "displayName": "Modified",
     "draft": false,
@@ -25,22 +25,22 @@ var expectedJson = {
     "isSite": false,
     "meta": {
         "mymodule:myschema": {
-            "a": ["1"]
+            "a": "1"
         },
         "mymodule:other": {
-            "name": ["test"]
+            "name": "test"
         }
     },
     "page": {
         "config": {
-            "a": ["1"]
+            "a": "1"
         },
         "controller": "mymodule:mycontroller",
         "regions": {
             "top": {
                 "components": [{
                     "config": {
-                        "a": ["1"]
+                        "a": "1"
                     },
                     "descriptor": "mymodule:mypart",
                     "name": "mypart",
@@ -48,7 +48,7 @@ var expectedJson = {
                     "type": "part"
                 }, {
                     "config": {
-                        "a": ["1"]
+                        "a": "1"
                     },
                     "descriptor": "mymodule:mylayout",
                     "name": "mylayout",
@@ -57,7 +57,7 @@ var expectedJson = {
                         "bottom": {
                             "components": [{
                                 "config": {
-                                    "a": ["1"]
+                                    "a": "1"
                                 },
                                 "descriptor": "mymodule:mypart",
                                 "name": "mypart",