+++ modules/portal-jslib/src/test/resources/com/enonic/xp/portal/jslib/impl/GetContentHandlerTest.js
@@ -7,32 +7,31 @@ var expectedJson = {
     "_name": "mycontent",
     "_path": "/a/b/mycontent",
     "data": {
-        "binaryReference": ["abc"],
+        "binaryReference": "abc",
         "boolean": [true, true, false],
-        "c": [{
-            "d": [true],
+        "c": {
+            "d": true,
             "e": ["3", "4", "5"],
-            "f": [2]
-        }],
-        "double": [2.2],
+            "f": 2
+        },
+        "double": 2.2,
         "doubles": [1.1, 2.2, 3.3],
-        "geoPoint": ["1.1,-1.1"],
+        "geoPoint": "1.1,-1.1",
         "geoPoints": ["1.1,-1.1", "2.2,-2.2"],
-        "htmlPart": ["<p>some<b>html</b></p>"],
-        "instant": ["+1000000000-12-31T23:59:59.999999999Z"],
-        "link": ["/my/content"],
-        "localDate": ["2014-01-31"],
-        "localDateTime": ["2014-01-31T10:30:05"],
-        "long": [1],
+        "htmlPart": "<p>some<b>html</b></p>",
+        "instant": "+1000000000-12-31T23:59:59.999999999Z",
+        "link": "/my/content",
+        "localDate": "2014-01-31",
+        "localDateTime": "2014-01-31T10:30:05",
+        "long": 1,
         "longs": [1, 2, 3],
-        "set": [{
-            "property": ["value"]
-        }],
-        "string": ["a"],
-        "stringEmpty": [""],
-        "stringNull": [],
+        "set": {
+            "property": "value"
+        },
+        "string": "a",
+        "stringEmpty": "",
         "strings": ["a", "b", "c"],
-        "xml": ["<xml><my-xml hello='world'/></xml>"]
+        "xml": "<xml><my-xml hello='world'/></xml>"
     },
     "displayName": "My Content",
     "draft": false,
@@ -41,19 +40,19 @@ var expectedJson = {
     "isSite": false,
     "meta": {
         "mymodule:myschema": {
-            "a": ["1"]
+            "a": "1"
         }
     },
     "page": {
         "config": {
-            "a": ["1"]
+            "a": "1"
         },
         "controller": "mymodule:mycontroller",
         "regions": {
             "top": {
                 "components": [{
                     "config": {
-                        "a": ["1"]
+                        "a": "1"
                     },
                     "descriptor": "mymodule:mypart",
                     "name": "mypart",
@@ -61,7 +60,7 @@ var expectedJson = {
                     "type": "part"
                 }, {
                     "config": {
-                        "a": ["1"]
+                        "a": "1"
                     },
                     "descriptor": "mymodule:mylayout",
                     "name": "mylayout",
@@ -70,7 +69,7 @@ var expectedJson = {
                         "bottom": {
                             "components": [{
                                 "config": {
-                                    "a": ["1"]
+                                    "a": "1"
                                 },
                                 "descriptor": "mymodule:mypart",
                                 "name": "mypart",