+++ CHANGELOG.md
@@ -10,11 +10,12 @@ Bugfixes:
 
 Features:
 
-  - N/A
+  - Simplify property-tree values in portal json result. Empty arrays is transformed to nul. Single value arrays
+    is transformed to a single value. (CMS-4924)
 
 Refactoring:
 
-  - N/A
+  - Removed general create URL method. Use pageUrl instead. (CMS-4979)
 
 
 ## RC2 (2015-01-20)